LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/resolve-export.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: Indirectly exported binding name 'f' is not found.

PASS Re-export of missing Wasm export should result in SyntaxError.
